- Home : could now swipe between tabs - Movement list: now display if a movement has been written in the blockchain - Add wallet: display as Dialog (DialogFragment) - UI Style (Navigation/ListView) : use round image with background color - cleaning code

+++ app/src/main/java/io/ucoin/app/service/AccountService.java
@@ -125,7 +125,7 @@ public class AccountService extends BaseService {
             wallet.setName(account.getUid());
             wallet.setCurrencyId(currency.getId());
             wallet.setAccountId(account.getId());
-            wallet.setCredit(credit == null ? 0 : credit.intValue());
+            wallet.setCredit(credit == null ? 0 : credit);
             WalletService walletService = ServiceLocator.instance().getWalletService();
             wallet = walletService.save(context, wallet);
         }
